Camera selection

+++ src/image.rs
@@ -10,7 +10,7 @@ use {vector::xy, image::Image};
 }
 
 pub fn upscale(target: &mut Image<&mut [u32]>, source: Image<&[u16]>) {
-    let [min, max] = [*source.iter().min().unwrap() as u16, *source.iter().max().unwrap() as u16];
+    let [min, max] = [*source.iter().min().unwrap(), *source.iter().max().unwrap()];
     if !(min < max) { return; }
     let [num, den] = if source.size.x*target.size.y > source.size.y*target.size.x { [target.size.x, source.size.x] } else { [target.size.y, source.size.y] };
     let target_size = source.size*(num/den); // largest integer fit
